Move `PerformanceCharge` to it's own file
