+++ Theatrical-Players-Refactoring-Kata/Performance.swift
@@ -2,10 +2,3 @@ struct Performance {
     let playID: String
     let audience: Int
 }
-
-struct PerformanceCharge {
-    let playName: String
-    let cost: Int
-    let volumeCredits: Int
-    let attendanceCount: Int
-}
