+++ Theatrical-Players-Refactoring-Kata/PerformanceCharge.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct PerformanceCharge {
+    let playName: String
+    let cost: Int
+    let volumeCredits: Int
+    let attendanceCount: Int
+}
